Fix iteration on Python 3.7
PEP 479 changed handling of StopIteration inside generators. Since
Python 3.7, StopIteration errors raised from generators are implicitly
converted to RuntimeErrors. The correct way of reporting the state
when there is nothing more to yield is to return.

Import from future ensures the same behavior on older versions of
Python that don’t have this as default.